requirements.txt
asgiref==3.7.2 ; python_version >= "3.10" and python_version < "3.11"
cffi==1.15.1 ; python_version >= "3.10" and python_version < "3.11"
cfgv==3.4.0 ; python_version >= "3.10" and python_version < "3.11"
colorama==0.4.6 ; python_version >= "3.10" and python_version < "3.11" and sys_platform == "win32"
coverage[toml]==7.3.0 ; python_version >= "3.10" and python_version < "3.11"
cryptography==41.0.3 ; python_version >= "3.10" and python_version < "3.11"
distlib==0.3.7 ; python_version >= "3.10" and python_version < "3.11"
django-cors-headers==4.2.0 ; python_version >= "3.10" and python_version < "3.11"
django-debug-toolbar==4.2.0 ; python_version >= "3.10" and python_version < "3.11"
django-extensions==3.2.3 ; python_version >= "3.10" and python_version < "3.11"
django-stubs-ext==4.2.2 ; python_version >= "3.10" and python_version < "3.11"
django-stubs[compatible-mypy]==4.2.4 ; python_version >= "3.10" and python_version < "3.11"
django==4.2.4 ; python_version >= "3.10" and python_version < "3.11"
djangorestframework-simplejwt[crypto]==5.2.2 ; python_version >= "3.10" and python_version < "3.11"
djangorestframework==3.14.0 ; python_version >= "3.10" and python_version < "3.11"
et-xmlfile==1.1.0 ; python_version >= "3.10" and python_version < "3.11"
exceptiongroup==1.1.3 ; python_version >= "3.10" and python_version < "3.11"
filelock==3.12.2 ; python_version >= "3.10" and python_version < "3.11"
freezegun==1.2.2 ; python_version >= "3.10" and python_version < "3.11"
identify==2.5.26 ; python_version >= "3.10" and python_version < "3.11"
iniconfig==2.0.0 ; python_version >= "3.10" and python_version < "3.11"
mypy-extensions==1.0.0 ; python_version >= "3.10" and python_version < "3.11"
mypy==1.5.1 ; python_version >= "3.10" and python_version < "3.11"
nodeenv==1.8.0 ; python_version >= "3.10" and python_version < "3.11"
numpy==1.25.2 ; python_version >= "3.10" and python_version < "3.11"
openpyxl==3.1.2 ; python_version >= "3.10" and python_version < "3.11"
packaging==23.1 ; python_version >= "3.10" and python_version < "3.11"
pandas==2.0.3 ; python_version >= "3.10" and python_version < "3.11"
platformdirs==3.10.0 ; python_version >= "3.10" and python_version < "3.11"
pluggy==1.2.0 ; python_version >= "3.10" and python_version < "3.11"
pre-commit==3.3.3 ; python_version >= "3.10" and python_version < "3.11"
psycopg2-binary==2.9.7 ; python_version >= "3.10" and python_version < "3.11"
pycparser==2.21 ; python_version >= "3.10" and python_version < "3.11"
pyjwt==2.8.0 ; python_version >= "3.10" and python_version < "3.11"
pytest-cov==4.1.0 ; python_version >= "3.10" and python_version < "3.11"
pytest-django==4.5.2 ; python_version >= "3.10" and python_version < "3.11"
pytest-env==1.0.1 ; python_version >= "3.10" and python_version < "3.11"
pytest-freezegun==0.4.2 ; python_version >= "3.10" and python_version < "3.11"
pytest==7.4.0 ; python_version >= "3.10" and python_version < "3.11"
python-dateutil==2.8.2 ; python_version >= "3.10" and python_version < "3.11"
pytz==2023.3 ; python_version >= "3.10" and python_version < "3.11"
pyyaml==6.0.1 ; python_version >= "3.10" and python_version < "3.11"
setuptools==68.1.2 ; python_version >= "3.10" and python_version < "3.11"
six==1.16.0 ; python_version >= "3.10" and python_version < "3.11"
sqlparse==0.4.4 ; python_version >= "3.10" and python_version < "3.11"
tomli==2.0.1 ; python_version >= "3.10" and python_version < "3.11"
types-pytz==2023.3.0.1 ; python_version >= "3.10" and python_version < "3.11"
types-pyyaml==6.0.12.11 ; python_version >= "3.10" and python_version < "3.11"
typing-extensions==4.7.1 ; python_version >= "3.10" and python_version < "3.11"
tzdata==2023.3 ; python_version >= "3.10" and python_version < "3.11"
virtualenv==20.24.3 ; python_version >= "3.10" and python_version < "3.11"
whitenoise==6.5.0 ; python_version >= "3.10" and python_version < "3.11"
#firebase-admin; python_version >= "3.10" and python_version < "3.11"